refactor(core): extract semantic vector sync (#1108)
